Move action helper methods into helper class
