+++ fastlane/lib/fastlane/actions/xcode_select.rb
@@ -18,30 +18,13 @@ module Fastlane
     #   environment variable into the full Developer content path.
     #
     class XcodeSelectAction < Action
-
-      def self.find_xcode(req)
-        req = Gem::Requirement.new(req.to_s)
-
-        Gem::Specification.find_by_name('xcode-install')
-        require 'xcode/install'
-
-        installer = XcodeInstall::Installer.new
-        installed = installer.installed_versions.reverse
-        installed.detect do |xcode|
-          req.satisfied_by? Gem::Version.new(xcode.version)
-        end
-      rescue Gem::LoadError
-        UI.error("The 'xcode-install' gem is needed to select Xcode by version number.")
-        Actions.verify_gem!('xcode-install') # This is awkward and destined to fail but is used to get the install help messages
-      end
-
       def self.run(params)
         xcode_path =
           case
           when params[:path]
             params[:path]
           when version = params[:version]
-            xcode = find_xcode(version)
+            xcode = Helper::XcodeSelectHelper.find_xcode(version)
             UI.user_error!("Cannot find an installed Xcode satisfying '#{version}'") if xcode.nil?
 
             UI.verbose("Found Xcode version #{xcode.version} at #{xcode.path} satisfying requirement #{version}")
@@ -73,7 +56,7 @@ module Fastlane
                                        conflict_block: proc do |value|
                                          UI.user_error!("You cannot specify 'path' and '#{value.key}' options at the same time")
                                        end,
-                                       verify_block: Verify.method(:path_exists)
+                                       verify_block: Helper::XcodeSelectHelper::Verify.method(:path_exists)
                                       ),
           FastlaneCore::ConfigItem.new(key: :version,
                                        env_name: "FL_XCODE_VERSION",
@@ -83,7 +66,7 @@ module Fastlane
                                        conflict_block: proc do |value|
                                          UI.user_error!("You cannot specify 'version' and '#{value.key}' options at the same time")
                                        end,
-                                       verify_block: Verify.method(:requirement)
+                                       verify_block: Helper::XcodeSelectHelper::Verify.method(:requirement)
                                       )
         ]
       end
@@ -91,19 +74,6 @@ module Fastlane
       def self.is_supported?(platform)
         [:ios, :mac].include? platform
       end
-
-      module Verify
-        def self.requirement(req)
-          Gem::Requirement.new(req.to_s)
-        rescue Gem::Requirement::BadRequirementError
-          UI.user_error!("The requirement '#{req}' is not a valid RubyGems style requirement")
-        end
-
-        def self.path_exists(path)
-          UI.user_error!("Path '#{path}' does not exist") unless Dir.exist?(path)
-        end
-      end
-
     end
   end
 end
